+++ src/main/java/com/codeit/donggrina/common/api/SearchFilter.java
@@ -1,4 +1,4 @@
-package com.codeit.donggrina.domain.growth_history.dto.request;
+package com.codeit.donggrina.common.api;
 
 import java.util.List;
